# docs/openapi.yaml
openapi: 3.0.0
info:
  contact:
    name: Coeeter
    url: https://github.com/coeeter/aniways
  description: An anime tracking and streaming platform API
  title: AniWays API
  version: 1.0.0
paths:
...
  "/anime/{id}/episodes":
    get:
      description: Get anime episodes
      parameters:
        - description: Anime ID
          in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/models.EpisodeResponse"
                type: array
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.ErrorResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.ErrorResponse"
      summary: Get anime episodes
      tags:
        - Episodes
  "/anime/{id}/episodes/{episodeID}/servers":
    get:
      description: Get episode servers
      parameters:
        - description: Anime ID
          in: path
          name: id
          required: true
          schema:
            type: string
        - description: Episode ID
          in: path
          name: episodeID
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/models.EpisodeServerResponse"
                type: array
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.ErrorResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.ErrorResponse"
      summary: Get episode servers
      tags:
        - Episodes
  "/anime/{id}/episodes/servers/{serverID}":
    get:
      description: Get episode stream data
      parameters:
        - description: Anime ID
          in: path
          name: id
          required: true
          schema:
            type: string
        - description: Server ID
          in: path
          name: serverID
          required: true
          schema:
            type: string
        - description: Server name
          in: query
          name: server
          required: true
          schema:
            type: string
        - description: Stream type
          in: query
          name: type
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.StreamingDataResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.ErrorResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.ErrorResponse"
      summary: Get episode stream data
      tags:
        - Episodes
...
servers:
  - url: http://localhost:8080
components:
...
  securitySchemes:
    cookieAuth:
      in: cookie
      name: aniways_session
      type: apiKey
  schemas:
...
    models.EpisodeResponse:
      properties:
        id:
          example: V1StGXR8Z5jdHi6B
          type: string
        isFiller:
          example: false
          type: boolean
        number:
          example: 1
          type: integer
        title:
          example: The Attack Titan
          type: string
      type: object
    models.EpisodeServerResponse:
      properties:
        serverId:
          example: V1StGXR8Z5jdHi6B
          type: string
        serverName:
          example: vidstreaming
          type: string
        type:
          example: sub
          type: string
      type: object
    models.ErrorResponse:
      properties:
        error:
          example: Invalid request
          type: string
      type: object
...
    models.SegmentResponse:
      properties:
        end:
          example: 180
          type: integer
        start:
          example: 90
          type: integer
      type: object
...
    models.StreamingDataResponse:
      properties:
        intro:
          $ref: "#/components/schemas/models.SegmentResponse"
        outro:
          $ref: "#/components/schemas/models.SegmentResponse"
        source:
          $ref: "#/components/schemas/models.StreamingSourceResponse"
        tracks:
          items:
            $ref: "#/components/schemas/models.TrackResponse"
          type: array
      type: object
    models.StreamingSourceResponse:
      properties:
        hls:
          example: https://example.com/stream.m3u8
          type: string
        iframe:
          example: https://example.com/embed/abc123
          type: string
        proxyHls:
          example: /proxy?p=encodedurl&s=hd
          type: string
      type: object
    models.TrackResponse:
      properties:
        default:
          example: true
          type: boolean
        kind:
          example: captions
          type: string
        label:
          example: English
          type: string
        raw:
          example: https://example.com/subtitles.vtt
          type: string
        url:
          example: /proxy?p=encodedurl&s=hd
          type: string
      type: object
...